// src/core/http/generated/models/admin-users/storeAdminUserRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Users
 * Filtered OpenAPI specification for Admin Users.
 * OpenAPI spec version: 0.0.1
 */
import type { StoreAdminUserRequestTenantsItem } from './storeAdminUserRequestTenantsItem.ts';

export interface StoreAdminUserRequest {
  /** @maxLength 255 */
  name: string;
  /** @maxLength 255 */
  email: string;
  /** @minLength 8 */
  password: string;
  is_global_admin?: boolean;
  tenants?: StoreAdminUserRequestTenantsItem[];
}