// src/core/http/generated/models/admin-users/updateAdminUserRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Users
 * Filtered OpenAPI specification for Admin Users.
 * OpenAPI spec version: 0.0.1
 */

export interface UpdateAdminUserRequest {
  /** @maxLength 255 */
  name?: string;
  /** @maxLength 255 */
  email?: string;
  /**
     * @minLength 8
     * @nullable
     */
  password?: string | null;
  is_global_admin?: boolean;
}