// src/core/http/generated/models/admin-tenant-users/storeAdminTenantUserRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Tenant Users
 * Filtered OpenAPI specification for Admin Tenant Users.
 * OpenAPI spec version: 0.0.1
 */

export interface StoreAdminTenantUserRequest {
  /** @nullable */
  user_id?: number | null;
  /** @maxLength 255 */
  name?: string;
  /** @maxLength 255 */
  email?: string;
  /** @minLength 8 */
  password?: string;
  /**
     * @maxLength 50
     * @nullable
     */
  role?: string | null;
  active?: boolean;
  /** @minLength 8 */
  password_confirmation?: string;
}