// src/core/http/generated/models/admin-tenant-users/updateAdminTenantUserRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Tenant Users
 * Filtered OpenAPI specification for Admin Tenant Users.
 * OpenAPI spec version: 0.0.1
 */

export interface UpdateAdminTenantUserRequest {
  /** @maxLength 255 */
  name?: string;
  /** @maxLength 255 */
  email?: string;
  /**
     * @minLength 8
     * @nullable
     */
  password?: string | null;
  /** @maxLength 50 */
  role?: string;
  active?: boolean;
  /**
     * @minLength 8
     * @nullable
     */
  password_confirmation?: string | null;
}